+++ src/instruments/buildSrc/igniter/tasks.mjs
@@ -49,9 +49,13 @@ const ecamPages = [
         path: 'SD/Pages/Press',
     },
     {
+        name: 'bleed-page',
+        path: 'SD/Pages/Bleed',
+    },
+  {
         name: 'status-page',
         path: 'SD/Pages/Status',
-    },
+  }
 ];
 
 export function getInputs() {